Publish: move payload generation to util
## Why
It is big and self-contained enough to be detached from the actions file. This makes it easier to find and edit.

If the util file grows, this can moved to its own file `util/publish-payload`

## Change
No functional changes; tried to keep 1:1 (i.e. just moving things around safely)
