+++ ui/modal/modalPublishPreview/view.jsx
@@ -17,8 +17,7 @@ import LbcSymbol from 'component/common/lbc-symbol';
 import ChannelThumbnail from 'component/channelThumbnail';
 import * as ICONS from 'constants/icons';
 import Icon from 'component/common/icon';
-import { NO_FILE } from 'redux/actions/publish';
-import { PAYWALL } from 'constants/publish';
+import { NO_FILE, PAYWALL } from 'constants/publish';
 import * as STRIPE from 'constants/stripe';
 import { TO_SECONDS } from 'util/stripe';
 import { removeInternalTags } from 'util/tags';
